feat: Add debugEnemyMasterDataId property to DungeonController.cs
This commit adds the debugEnemyMasterDataId property to the DungeonController.cs file. The property allows for specifying the master data ID of the enemy character for debugging purposes. This enhancement improves the debugging capabilities of the game by providing a way to easily test specific enemy encounters.

Note: The commit message has been generated based on the provided code changes and recent commits.