feat: Update AcquireItemView to display item thumbnail
This commit updates the AcquireItemView class to display the thumbnail of the acquired item. It adds a line of code to set the sprite of the "Image" component to the thumbnail of the master data item. This enhancement improves the visual representation of the acquired item in the UI.

Note: The commit message has been generated based on the provided code changes and recent commits.