feat: Update CenterRadiationShaderCanvas Variant.mat
This commit updates the CenterRadiationShaderCanvas Variant.mat file in the Radiation/Shader directory of the EasyCustomVFXUI/Effect/1SpriteEffect folder. The changes include modifying the values of _CenterRadius, _Speed, _WholeRadius, and _Color properties. These modifications enhance the visual effects of the radiation shader in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.