feat: Update sfxName in GiveDamage.cs
This commit updates the sfxName variable in the GiveDamage.cs file. The previous value was "Sfx.Message.1" and it has been changed to "Sfx.Message.2". This change enhances the audio experience in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.