feat: Update enemy name and thumbnail in MasterData.asset
This commit updates the enemy name and thumbnail in the MasterData.asset file. The name of the enemy with Id 10105 has been changed to "天魔、ウォロト". Additionally, the thumbnail for this enemy has been updated. These changes enhance the accuracy and visual representation of the enemy data.

Note: The commit message has been generated based on the provided code changes and recent commits.