feat: Update AcquireItemView to display item name
This commit updates the AcquireItemView script to display the name of the acquired item. The masterDataItem's name is now assigned to the TMP_Text component with the "Name" tag. This enhancement improves the user experience by providing more information about the acquired item.

Note: The commit message has been generated based on the provided code changes and recent commits.