feat: Add new Effects assets for Battle.Begin sequence
This commit adds new Effects assets for the Battle.Begin sequence in the SoulRPG project. The Effects.meta and Battle.Begin.0.asset files have been created and added to the Effects folder. These assets define the visual effects used during the beginning of a battle. This addition enhances the visual experience and immersion in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.