feat: Add armor setup for different armor types in GameItemInformationView.cs
This commit adds the ability to set up armor information for different armor types in the GameItemInformationView.cs file. The code changes introduce conditional statements to check the item's ID and retrieve the corresponding armor data. This enhancement improves the display of armor information in the game's UI.

Note: The commit message has been generated based on the provided code changes and recent commits.