feat: Add PlaySfx sequence to trigger sound effects
This commit adds a new PlaySfx sequence to the project. The sequence is responsible for playing a specified sound effect (SFX) by broadcasting an event to the GameEvents system. This modification enhances the audio experience in the game.

Note: The commit message has been generated based on the provided code changes and recent commits.