feat: Update OnRequestShowMessage.cs to set default value for sfxName
This commit updates the OnRequestShowMessage.cs file to set a default value for the sfxName field. The default value is "Sfx.Message.0". This change ensures that the sfxName field is always initialized, improving the reliability of the code.

Note: The commit message has been generated based on the provided code changes and recent commits.