feat: Update anchored position of UI.Game.InputGuide.Center.prefab
This commit updates the anchored position of the UI.Game.InputGuide.Center.prefab in the Assets/SoulRPG/Prefabs/UI.Game.InputGuide.Center.prefab file. The y-coordinate of the m_AnchoredPosition property has been changed from 320 to 290. This modification improves the alignment of the input guide center in the game UI.

Note: The commit message has been generated based on the provided code changes and recent commits.