feat: Add GetTag extension method to InputAction
This commit adds the GetTag extension method to the InputAction class in the Extensions.Input.cs file. The method retrieves the tag associated with the input action using the InputSprite.GetTag method and the resolved InputScheme. This addition enhances the functionality of the InputAction class by providing an easy way to get the tag for a specific input action.

Note: The commit message has been generated based on the provided code changes and recent commits.